Move clear and view downloads to utils
