+++ blobfish/utils/cloud_utils.py
@@ -1,0 +1,29 @@
+import os
+import boto3
+
+def clear_downloads(bucket: str, prefix: str):
+    client = boto3.client(
+        service_name="s3",
+        aws_access_key_id=os.environ["AWS_ACCESS_KEY_ID"],
+        aws_secret_access_key=os.environ["AWS_SECRET_ACCESS_KEY"],
+        region_name=os.environ["AWS_DEFAULT_REGION"],
+    )
+    paginator = client.get_paginator("list_objects_v2")
+    for page in paginator.paginate(Bucket=bucket, Prefix=prefix):
+        contents = page.get("Contents", [])
+        for content in contents:
+            client.delete_object(Bucket=bucket, Key=content.get("Key"))
+
+
+def view_downloads(bucket: str, prefix: str):
+    client = boto3.client(
+        service_name="s3",
+        aws_access_key_id=os.environ["AWS_ACCESS_KEY_ID"],
+        aws_secret_access_key=os.environ["AWS_SECRET_ACCESS_KEY"],
+        region_name=os.environ["AWS_DEFAULT_REGION"],
+    )
+    paginator = client.get_paginator("list_objects_v2")
+    for page in paginator.paginate(Bucket=bucket, Prefix=prefix):
+        contents = page.get("Contents", [])
+        for content in contents:
+            print(client.get_object(Bucket=bucket, Key=content.get("Key")))
